Commit 068
- Implemented new Basic AAF Stat systems to be used in staged anims:
There will be a new bar meter called "WOOF". This meter will control which anim will play in a staged sequence, so for example:
a staged sequence with 4 anims has the first anim tagged as "Tease" and the last one as "Climax". The tease anim will be considered as foreplay and will only play until the woof bar gets to 20%. When the woof bar gets to 20% it'll play the intermediate anims until it gets to 80%, once at 80% the staged sequence will loop the cum animation until the woof stats gets back to 0 and the stage sequence is restarted.
Normally a staged sequence will last 90s from the starting stage to the climax stage, but you can choose in FOMOD how long you would like to have the staged sequence to last for staged sequences (60s, 90s, 120s or 180s).

- Implemented a initial narration system. Actors may say something at certain situations (like when WOOF is at 70% and is about to trigger the climax anim)

- During FOMOd installtion, you'll be able to choose if you want the staged sequences to play intermediate stages sequentially or randomly

- Incorporated the initial implementations for actors to use the cum particle effects in almost all climax anims

- Changed how often the manly moans topics will be triggered a bit

- Few changes to the morphsetData as well as adding the Anus Spread, Penis Adjust and Foreskin morphs to force them to 0 so nothing goes wrong during AAF anims

- Added penis Adjust morph to all creatures for the alignment project in gay anims

- Cleaned up the target atrribute from manly moan actions in all animationData.xmls, since topics can be trigered with self node alone

- Changed how rottweiler replacer addon is gonna work: Now, instead of overwriting the original dogmeat nif file, there will be a ARMA record override using a plugin (to make updates and tests easier). The old plugin for rottweiler replaces was renamed to rottweiler replacer.esp so be aware when you install this new version