Commit 097
positionTrees are handled differently in some mods, with that in mind. Some are npc controlled which causes branches to be automatically navigated. With that in mind:

- Set all positionTrees with useConditions="True" and strictExiting="True". So the branches will only advance based on WOOF stats + AAF will only be terminated automatically when a exit branch is reached (mostly a orgasm/climax anim).
In addition, all branches were set with a certain ammount of time so the automatic navigation wont delay too much between one branch and another (eg: when a SceneSetting is defined by a AAF mod).
Keep in mind, this will lock branches under certain WOOF stats range if the positionTree is called in a non NPCControlled interaction. So you can always reset the stats by changing between one positionTree and another
These settings have been done to both MM and FM anims from BP70 and Leito Animation Pack

- Resetall Action will work differently: Instead of applying a reaction set, it'll add -100

- Fixed issue with [UAP] Leito - Mattress Reverse Cowgirl 2 Tree referencing the [UAP] Leito - Mattress Reverse Cowgirl staged sequence instead of [UAP] Leito - Mattress Reverse Cowgirl 2 in the first branch